requirements.txt
Flask==2.3.3
pynmea2==1.19.0
georinex==1.13.0
pyproj==3.6.1
psycopg2-binary==2.9.9
pymongo==4.6.1
pandas==2.1.1
numpy==1.24.3
d3==0.0.2
python-dotenv==1.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
GeoAlchemy2==0.14.3
geopandas==0.14.1
Flask-Login==0.6.3
gunicorn==21.2.0
pytest==7.4.2
